+++ cakephp/app/Controller/TracksController.php
@@ -67,17 +67,23 @@ class TracksController extends AppController {
 	}
 	
 	function delete($id) {
-		if ($this->request->is('delete') && !empty($this->request->data)) {
-			if ($this->request->data['confirm']) {
-				if ($this->Track->delete($this->request->data['t_TrackID'])) {
-					$this->Session->setFlash('The track was successfully deleted.');
-					$this->redirect(array('controller' => 'musiclibrary', 'action' => 'index'));
+		$track = $this->Track->find('first', array('conditions' => array('t_TrackID' => $id)));
+		if (count($track['FloatingShowElement']) > 0) {
+			$this->set('track', $track);
+			$this->set('deleteDenied', true);
+		} else {
+			if ($this->request->is('delete') && !empty($this->request->data)) {
+				if ($this->request->data['confirm']) {
+					if ($this->Track->delete($this->request->data['t_TrackID'])) {
+						$this->Session->setFlash('The track was successfully deleted.');
+						$this->redirect(array('controller' => 'musiclibrary', 'action' => 'index'));
+					}
+				} else {
+					$this->redirect(array('controller' => 'tracks', 'action' => 'view', $this->request->data['t_TrackID']));
 				}
 			} else {
-				$this->redirect(array('controller' => 'tracks', 'action' => 'view', $this->request->data['t_TrackID']));
+				$this->set('trackId', $id);
 			}
-		} else {
-			$this->set('trackId', $id);
 		}
 	}
